Add support for review metrics and bounds browsing
* Refactor review methods into it's own dal/controller

* Add getReviewMetrics

* Add getReviewTaskClusters

* Add location (boundingBox) searching to review methods
