+++ app/org/maproulette/controllers/api/TaskController.scala
@@ -226,157 +226,6 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
   }
 
   /**
-    * Gets and claims a task that needs to be reviewed.
-    *
-    * @param id Task id to work on
-    * @return
-    */
-  def startTaskReview(id:Long) : Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val task = this.dal.retrieveById(id) match {
-        case Some(t) => t
-        case None => throw new NotFoundException(s"Task with $id not found, cannot start review.")
-      }
-
-      val result = this.dal.startTaskReview(user, task)
-      Ok(Json.toJson(result))
-    }
-  }
-
-  /**
-    * Releases a claim on a task that needs to be reviewed.
-    *
-    * @param id Task id to work on
-    * @return
-    */
-  def cancelTaskReview(id:Long) : Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val task = this.dal.retrieveById(id) match {
-        case Some(t) => t
-        case None => throw new NotFoundException(s"Task with $id not found, cannot cancel review.")
-      }
-
-      val result = this.dal.cancelTaskReview(user, task)
-      Ok(Json.toJson(result))
-    }
-  }
-
-  /**
-    * Gets and claims the next task that needs to be reviewed.
-    *
-    * Valid search parameters include:
-    * cs => "my challenge name"
-    * o => "mapper's name"
-    * r => "reviewer's name"
-    *
-    * @return Task
-    */
-  def nextTaskReview(sort:String, order:String) : Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      SearchParameters.withSearch { implicit params =>
-        val result = this.dal.nextTaskReview(user, params, sort, order)
-        val nextTask = result match {
-          case Some(task) =>
-            Ok(Json.toJson(this.dal.startTaskReview(user, task)))
-          case None =>
-            throw new NotFoundException("No tasks found to review.")
-        }
-
-        nextTask
-      }
-    }
-  }
-
-  /**
-    * Gets tasks where a review is requested
-    *
-    * @param startDate Optional start date to filter by reviewedAt date
-    * @param endDate Optional end date to filter by reviewedAt date
-    * @param limit The number of tasks to return
-    * @param page The page number for the results
-    * @param sort The column to sort
-    * @param order The order direction to sort
-    * @return
-    */
-  def getReviewRequestedTasks(startDate: String=null, endDate: String=null,
-                              limit:Int, page:Int, sort:String, order:String) : Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      SearchParameters.withSearch { implicit params =>
-        //cs => "my challenge name"
-        //o => "mapper's name"
-        //r => "reviewer's name"
-        val (count, result) = this.dal.getReviewRequestedTasks(User.userOrMocked(user), params,
-           startDate, endDate, limit, page, sort, order)
-        Ok(Json.obj("total" -> count, "tasks" -> _insertExtraJSON(result)))
-      }
-    }
-  }
-
-  /**
-    * Gets reviewed tasks where the user has reviewed or requested review
-    *
-    * @param startDate Optional start date to filter by reviewedAt date
-    * @param endDate Optional end date to filter by reviewedAt date
-    * @param asReviewer Whether we should return tasks reviewed by this user or reqested by this user
-    * @param allowReviewNeeded Whether we should return tasks where status is review requested also
-    * @param limit The number of tasks to return
-    * @param page The page number for the results
-    * @param sort The column to sort
-    * @param order The order direction to sort
-    * @return
-    */
-  def getReviewedTasks(startDate: String=null, endDate: String=null,
-                       asReviewer: Boolean=false, allowReviewNeeded: Boolean=false,
-                       limit:Int, page:Int,
-                       sort:String, order:String) : Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      SearchParameters.withSearch { implicit params =>
-        val (count, result) = this.dal.getReviewedTasks(User.userOrMocked(user), params,
-             startDate, endDate, asReviewer, allowReviewNeeded, limit, page, sort, order)
-        Ok(Json.obj("total" -> count, "tasks" -> _insertExtraJSON(result)))
-      }
-    }
-  }
-
-  /**
-   * Fetches the matching parent object and inserts it (id, name, status)
-   * into the JSON data returned. Also fetches and inserts usernames for
-   * 'reviewRequestedBy' and 'reviewBy'
-   */
-  private def _insertExtraJSON(tasks: List[Task]): JsValue = {
-    if (tasks.isEmpty) {
-      Json.toJson(List[JsValue]())
-    } else {
-      val challenges = Some(this.dalManager.challenge.retrieveListById(-1, 0)(tasks.map(c =>
-        c.parent)).map(c => c.id -> Json.obj("id" -> c.id, "name" -> c.name, "status" -> c.status)).toMap)
-
-      val mappers = Some(this.dalManager.user.retrieveListById(-1, 0)(tasks.map(
-        t => t.reviewRequestedBy.getOrElse(0L))).map(u =>
-          u.id -> Json.obj("username" -> u.name, "id" -> u.id)).toMap)
-
-      val reviewers = Some(this.dalManager.user.retrieveListById(-1, 0)(tasks.map(
-        t => t.reviewedBy.getOrElse(0L))).map(u =>
-          u.id -> Json.obj("username" -> u.name, "id" -> u.id)).toMap)
-
-      val jsonList = tasks.map { task =>
-        val challengeJson = Json.toJson(challenges.get(task.parent)).as[JsObject]
-        var updated = Utils.insertIntoJson(Json.toJson(task), Challenge.KEY_PARENT, challengeJson, true)
-        if (task.reviewRequestedBy.getOrElse(0) != 0) {
-          val mapperJson = Json.toJson(mappers.get(task.reviewRequestedBy.get)).as[JsObject]
-          updated = Utils.insertIntoJson(updated, "reviewRequestedBy", mapperJson, true)
-        }
-        if (task.reviewedBy.getOrElse(0) != 0) {
-          val reviewerJson = Json.toJson(reviewers.get(task.reviewedBy.get)).as[JsObject]
-          updated = Utils.insertIntoJson(updated, "reviewedBy", reviewerJson, true)
-        }
-
-        updated
-      }
-      Json.toJson(jsonList)
-    }
-  }
-
-  /**
     * Gets all the tasks within a bounding box
     *
     * @param left   The minimum latitude for the bounding box
