+++ app/org/maproulette/models/dal/TaskDAL.scala
@@ -529,6 +529,24 @@ class TaskDAL @Inject()(override val db: Database,
     }
   }
 
+  def getTaskWithReview(taskId: Long): TaskWithReview = {
+    this.withMRConnection { implicit c =>
+      val query = s"""
+        SELECT $retrieveColumnsWithReview,
+               challenges.name as challenge_name,
+               mappers.name as review_requested_by_username,
+               reviewers.name as reviewed_by_username
+        FROM ${this.tableName}
+        LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id
+        LEFT OUTER JOIN users mappers ON task_review.review_requested_by = mappers.id
+        LEFT OUTER JOIN users reviewers ON task_review.reviewed_by = reviewers.id
+        INNER JOIN challenges ON challenges.id = tasks.parent_id
+        WHERE tasks.id = {taskId}
+      """
+      SQL(query).on('taskId -> taskId).as(this.taskWithReviewParser.single)
+    }
+  }
+
   /**
     * Sets the task for a given user. The user cannot set the status of a task unless the object has
     * been locked by the same user before hand.
@@ -1212,348 +1230,6 @@ class TaskDAL @Inject()(override val db: Database,
   }
 
   /**
-    * Gets and claims a task for review.
-    *
-    * @param user The user executing the request
-    * @param id id of task that you wish to start/claim
-    * @return task
-    */
-  def startTaskReview(user:User, task:Task) (implicit c:Connection=null) : Option[Task] = {
-    if (task.reviewClaimedBy.getOrElse(null) != null &&
-        task.reviewClaimedBy.getOrElse(null) != user.id.toLong) {
-      throw new InvalidException("This task is already being reviewed by someone else.")
-    }
-
-    this.withMRTransaction { implicit c =>
-      // Unclaim everything before starting a new task.
-      SQL"""UPDATE task_review SET review_claimed_by = NULL, review_claimed_at = NULL
-              WHERE review_claimed_by = #${user.id}""".executeUpdate()
-
-      val updatedRows =
-        SQL"""UPDATE task_review SET review_claimed_by = #${user.id}, review_claimed_at = NOW()
-                WHERE task_id = #${task.id} AND review_claimed_at IS NULL""".executeUpdate()
-
-      // if returning 0, then this is because the item is locked by a different user
-      if (updatedRows == 0) {
-        throw new IllegalAccessException(s"Current task [${task.id} is locked by another user, cannot start review at this time.")
-      }
-
-      try {
-        this.unlockItem(user, task)
-      } catch {
-        case e: Exception => logger.warn(e.getMessage)
-      }
-
-      webSocketProvider.sendMessage(WebSocketMessages.reviewClaimed(
-        WebSocketMessages.ReviewData(this.getTaskWithReview(task.id))
-      ))
-    }
-
-    val updatedTask = task.copy(reviewClaimedBy = Option(user.id.toInt))
-    this.cacheManager.withOptionCaching { () => Some(updatedTask) }
-    Option(updatedTask)
-  }
-
-  /**
-    * Releases a claim on a task for review.
-    *
-    * @param user The user executing the request
-    * @param id id of task that you wish to release
-    * @return task
-    */
-  def cancelTaskReview(user:User, task:Task) (implicit c:Connection=null) : Option[Task] = {
-    if (task.reviewClaimedBy.getOrElse(null) != user.id.toLong) {
-      throw new InvalidException("This task is not currently being reviewed by you.")
-    }
-
-    this.withMRTransaction { implicit c =>
-      val updatedRows = SQL"""UPDATE task_review t SET review_claimed_by = NULL, review_claimed_at = NULL
-              WHERE t.task_id = #${task.id}""".executeUpdate()
-
-      // if returning 0, then this is because the item is locked by a different user
-      if (updatedRows == 0) {
-        throw new IllegalAccessException(s"Current task [${task.id} is locked by another user, cannot cancel review at this time.")
-      }
-
-      webSocketProvider.sendMessage(WebSocketMessages.reviewUpdate(
-        WebSocketMessages.ReviewData(this.getTaskWithReview(task.id))
-      ))
-    }
-
-    try {
-      this.unlockItem(user, task)
-    } catch {
-      case e: Exception => logger.warn(e.getMessage)
-    }
-
-    val updatedTask = task.copy(reviewClaimedBy = None)
-    this.cacheManager.withOptionCaching { () => Some(updatedTask) }
-    Option(updatedTask)
-  }
-
-  /**
-    * Gets and claims the next task for review.
-    *
-    * @param user The user executing the request
-    * @return task
-    */
-  def nextTaskReview(user:User, searchParameters: SearchParameters,
-                    sort:String, order:String) (implicit c:Connection=null) : Option[Task] = {
-    val (count, result) = this.getReviewRequestedTasks(user, searchParameters, null, null, 1, 0, sort, order, false)
-    if (count == 0) {
-      return None
-    }
-    else {
-      return Some(result.head)
-    }
-  }
-
-  def getTaskWithReview(taskId: Long): TaskWithReview = {
-    this.withMRConnection { implicit c =>
-      val query = s"""
-        SELECT $retrieveColumnsWithReview,
-               challenges.name as challenge_name,
-               mappers.name as review_requested_by_username,
-               reviewers.name as reviewed_by_username
-        FROM ${this.tableName}
-        LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id
-        LEFT OUTER JOIN users mappers ON task_review.review_requested_by = mappers.id
-        LEFT OUTER JOIN users reviewers ON task_review.reviewed_by = reviewers.id
-        INNER JOIN challenges ON challenges.id = tasks.parent_id
-        WHERE tasks.id = {taskId}
-      """
-      SQL(query).on('taskId -> taskId).as(this.taskWithReviewParser.single)
-    }
-  }
-
-  /**
-    * Gets a list of tasks that have requested review (and are in this user's project group)
-    *
-    * @param user The user executing the request
-    * @param startDate Limit tasks to reviewed after date (YYYY-MM-DD)
-    * @param endDate Limit tasks to reviewed before date (YYYY-MM-DD)
-    * @param limit The number of tasks to return
-    * @param offset Offset to start paging
-    * @param sort Sort column
-    * @param order DESC or ASC
-    * @param includeDisputed Whether disputed tasks whould be returned in results (default is true)
-    * @return A list of tasks
-    */
-  def getReviewRequestedTasks(user:User, searchParameters: SearchParameters,
-                              startDate:String, endDate:String,
-                              limit:Int = -1, offset:Int=0, sort:String, order:String, includeDisputed: Boolean = true)
-                    (implicit c:Connection=null) : (Int, List[Task]) = {
-    var orderByClause = ""
-    val whereClause = includeDisputed match {
-      case true =>
-        new StringBuilder(s"(task_review.review_status=${Task.REVIEW_STATUS_REQUESTED} OR task_review.review_status=${Task.REVIEW_STATUS_DISPUTED})")
-      case false =>
-        new StringBuilder(s"task_review.review_status=${Task.REVIEW_STATUS_REQUESTED}")
-    }
-
-    val joinClause = new StringBuilder("INNER JOIN challenges c ON c.id = tasks.parent_id ")
-    joinClause ++= "LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id "
-
-    this.appendInWhereClause(whereClause,
-      s"(task_review.review_claimed_at IS NULL OR task_review.review_claimed_by = ${user.id})")
-
-    val parameters = new ListBuffer[NamedParameter]()
-    parameters ++= addSearchToQuery(searchParameters, whereClause)
-
-    setupReviewSearchClause(whereClause, joinClause, searchParameters, startDate, endDate)
-
-    sort match {
-      case s if s.nonEmpty =>
-        orderByClause = this.order(Some(s), order, "", false)
-      case _ => // ignore
-    }
-
-    val query = user.isSuperUser match {
-      case true =>
-        s"""
-          SELECT tasks.${this.retrieveColumnsWithReview} FROM tasks
-          ${joinClause}
-          WHERE
-          ${whereClause}
-          ${orderByClause}
-          LIMIT ${sqlLimit(limit)} OFFSET ${offset}
-         """
-      case default =>
-        if (user.settings.isReviewer.getOrElse(false)) {
-          // You see review task when:
-          // 1. Project and Challenge enabled (so visible to everyone)
-          // 2. You own the Project
-          // 3. You manage the project (your user group matches groups of project)
-          s"""
-            SELECT tasks.${this.retrieveColumnsWithReview} FROM tasks
-            ${joinClause}
-            INNER JOIN projects p ON p.id = c.parent_id
-            INNER JOIN groups g ON g.project_id = p.id
-            INNER JOIN user_groups ug ON g.id = ug.group_id
-            WHERE ((p.enabled AND c.enabled) OR
-                    p.owner_id = ${user.osmProfile.id} OR
-                    ug.osm_user_id = ${user.osmProfile.id}) AND
-                    task_review.review_requested_by != ${user.id} AND
-            ${whereClause}
-            ${orderByClause}
-            LIMIT ${sqlLimit(limit)} OFFSET ${offset}
-           """
-         }
-         else {
-           return (0, List[Task]())
-         }
-    }
-
-    val countQuery = user.isSuperUser match {
-      case true =>
-        s"""
-          SELECT count(*) FROM tasks
-          ${joinClause}
-          WHERE ${whereClause}
-        """
-      case default =>
-        s"""
-          SELECT count(*) FROM tasks
-          ${joinClause}
-          INNER JOIN projects p ON p.id = c.parent_id
-          INNER JOIN groups g ON g.project_id = p.id
-          INNER JOIN user_groups ug ON g.id = ug.group_id
-          WHERE ((p.enabled AND c.enabled) OR
-                  p.owner_id = ${user.osmProfile.id} OR
-                  ug.osm_user_id = ${user.osmProfile.id}) AND
-                  task_review.review_requested_by != ${user.id} AND
-          ${whereClause}
-        """
-    }
-
-    var count = 0
-    val tasks = this.cacheManager.withIDListCaching { implicit cachedItems =>
-      this.withMRTransaction { implicit c =>
-        count = sqlWithParameters(countQuery, parameters).as(SqlParser.int("count").single)
-        sqlWithParameters(query, parameters).as(this.parser.*)
-      }
-    }
-
-    return (count, tasks)
-  }
-
-  /**
-    * Gets a list of tasks that have been reviewed (either by this user or requested by this user)
-    *
-    * @param user The user executing the request
-    * @param startDate Limit tasks to reviewed after date (YYYY-MM-DD)
-    * @param endDate Limit tasks to reviewed before date (YYYY-MM-DD)
-    * @param asReviewer Whether we should return tasks reviewed by this user or reqested by this user
-    * @param allowReviewNeeded Whether we should include review requested tasks as well
-    * @param limit The amount of tasks to be returned
-    * @param offset Offset to start paging
-    * @param sort Column to sort
-    * @param order DESC or ASC
-    * @return A list of tasks
-    */
-  def getReviewedTasks(user:User, searchParameters: SearchParameters,
-                       startDate:String, endDate:String,
-                       asReviewer:Boolean=false, allowReviewNeeded:Boolean=false,
-                       limit:Int = -1, offset:Int=0, sort:String, order:String)
-                       (implicit c:Connection=null) : (Int, List[Task]) = {
-    val fetchBy = if (asReviewer) "task_review.reviewed_by" else "task_review.review_requested_by"
-    var orderByClause = ""
-    val whereClause = new StringBuilder(s"${fetchBy}=${user.id}")
-    val joinClause = new StringBuilder("INNER JOIN challenges c ON c.id = tasks.parent_id ")
-    joinClause ++= "LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id "
-
-    val parameters = new ListBuffer[NamedParameter]()
-    parameters ++= addSearchToQuery(searchParameters, whereClause)
-
-    if (!allowReviewNeeded) {
-      this.appendInWhereClause(whereClause, s"task_review.review_status <> ${Task.REVIEW_STATUS_REQUESTED} ")
-    }
-
-    setupReviewSearchClause(whereClause, joinClause, searchParameters, startDate, endDate)
-
-    sort match {
-      case s if s.nonEmpty =>
-        orderByClause = this.order(Some(s), order, "", false)
-      case _ => // ignore
-    }
-
-    val countQuery = s"""
-     SELECT count(*) FROM tasks
-     ${joinClause}
-     WHERE
-     ${whereClause}
-    """
-
-    val query = s"""
-     SELECT tasks.${this.retrieveColumnsWithReview} FROM tasks
-     ${joinClause}
-     WHERE
-     ${whereClause}
-     ${orderByClause}
-     LIMIT ${sqlLimit(limit)} OFFSET $offset
-    """
-
-   var count = 0
-   val tasks = this.cacheManager.withIDListCaching { implicit cachedItems =>
-      this.withMRTransaction { implicit c =>
-        count = sqlWithParameters(countQuery, parameters).as(SqlParser.int("count").single)
-        sqlWithParameters(query, parameters).as(this.parser.*)
-      }
-    }
-
-    return (count, tasks)
-  }
-
-  /**
-   * private setup the search clauses for searching the review tables
-   */
-  private def setupReviewSearchClause(whereClause: StringBuilder, joinClause: StringBuilder,
-                                      searchParameters: SearchParameters,
-                                      startDate: String, endDate: String) {
-    searchParameters.owner match {
-     case Some(o) if o.nonEmpty =>
-       joinClause ++= "INNER JOIN users u ON u.id = task_review.review_requested_by "
-       this.appendInWhereClause(whereClause, s"LOWER(u.name) LIKE LOWER('%${o}%')")
-     case _ => // ignore
-    }
-
-    searchParameters.reviewer match {
-     case Some(r) if r.nonEmpty =>
-       joinClause ++= "INNER JOIN users u2 ON u2.id = task_review.reviewed_by "
-       this.appendInWhereClause(whereClause, s"LOWER(u2.name) LIKE LOWER('%${r}%')")
-     case _ => // ignore
-    }
-
-    searchParameters.taskStatus match {
-      case Some(statuses) if statuses.nonEmpty =>
-        val statusClause = new StringBuilder(s"(tasks.status IN (${statuses.mkString(",")})")
-        if (statuses.contains(-1)) {
-          statusClause ++= " OR c.status IS NULL"
-        }
-        statusClause ++= ")"
-        this.appendInWhereClause(whereClause, statusClause.toString())
-      case Some(statuses) if statuses.isEmpty => //ignore this scenario
-      case _ =>
-    }
-
-    searchParameters.taskReviewStatus match {
-      case Some(statuses) if statuses.nonEmpty =>
-        val statusClause = new StringBuilder(s"(task_review.review_status IN (${statuses.mkString(",")}))")
-        this.appendInWhereClause(whereClause, statusClause.toString())
-      case Some(statuses) if statuses.isEmpty => //ignore this scenario
-      case _ =>
-    }
-
-    if (startDate != null && startDate.matches("[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]")) {
-      this.appendInWhereClause(whereClause, "reviewed_at >= '" + startDate + " 00:00:00'")
-    }
-
-    if (endDate != null && endDate.matches("[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]")) {
-      this.appendInWhereClause(whereClause, "reviewed_at <= '" + endDate + " 23:59:59'")
-    }
-  }
-
-  /**
     * This function will retrieve all the tasks in a given bounded area. You can use various search
     * parameters to limit the tasks retrieved in the bounding box area.
     *
@@ -1821,5 +1497,5 @@ class TaskDAL @Inject()(override val db: Database,
 }
 
 object TaskDAL {
-  private val DEFAULT_NUMBER_OF_POINTS = 100
+  val DEFAULT_NUMBER_OF_POINTS = 100
 }
